+++ app/controllers/tasks_controller.rb
@@ -14,6 +14,10 @@ class TasksController < ApplicationController
       participations_only: false
     ).objects
 
+    # serialize using the base task serializer which acts as a lightweight
+    # summary of each task rather than each custom task serializer that may
+    # side load an excessive amount of data that is unnecessary for an index
+    # response.
     respond_with tasks, each_serializer: TaskSerializer
   end
 
@@ -126,8 +130,23 @@ class TasksController < ApplicationController
   end
 
   def new_task_params
-    paper = Paper.find_by_id_or_short_doi(params[:task][:paper_id])
-    task_params(task_type).merge(paper: paper, creator: paper.creator)
+    task_params(task_type).dup.tap do |new_params|
+      # for all new tasks, assign necessary paper attributes
+      new_params[:paper] = paper
+      new_params[:creator] = paper.creator
+
+      # for custom cards, assign the latest card version
+      if params[:task][:type] == 'CustomCardTask'
+        new_params[:card_version] = card_version
+      end
+    end
+  end
+
+  def card_version
+    @card_version ||= begin
+      card = paper.journal.cards.find_by(id: params[:task][:card_id])
+      card.card_version(:latest)
+    end
   end
 
   def unmunge_empty_arrays