+++ app/models/card_content.rb
@@ -4,6 +4,8 @@
 # text, or widgets (developer-created chunks of functionality with
 # user-configured behavior)
 class CardContent < ActiveRecord::Base
+  include XmlSerializable
+
   acts_as_nested_set
   acts_as_paranoid
 
@@ -25,6 +27,43 @@ class CardContent < ActiveRecord::Base
               message: "CardContent idents must be unique"
             },
             if: -> { ident.present? }
+
+  validate :content_value_type_combination
+
+  SUPPORTED_VALUE_TYPES = %w(attachment boolean question-set text html).freeze
+
+  # Note that value_type really refers to the value_type of answers associated
+  # with this piece of card content. In the old NestedQuestion world, both
+  # NestedQuestionAnswer and NestedQuestion had a value_type column, and the
+  # value_type was duplicated between them. In the hash below, we say that the
+  # 'short-input' answers will have a 'text' value type, while 'radio' answers
+  # can either be boolean or text.  The 'text' content_type is really static
+  # text, which will never have an answer associated with it, hence it has no
+  # possible value types.  The same goes for the other container types
+  # (field-set, etc)
+  VALUE_TYPES_FOR_CONTENT =
+    { 'display-children': [nil],
+      'display-with-value': [nil],
+      'field-set': [nil],
+      'short-input': ['text'],
+      'check-box': ['boolean'],
+      'text': [nil],
+      'paragraph-input': ['text', 'html'],
+      'radio': ['boolean', 'text'] }.freeze.with_indifferent_access
+
+  # Although we want to validate the various combinations of content types
+  # and value types, many of the CardContent records that have been created
+  # via the CardLoader don't have a content_type set at all, so we'll skip
+  # validating those
+  def content_value_type_combination
+    return if content_type.blank?
+    unless VALUE_TYPES_FOR_CONTENT.fetch(content_type, []).member?(value_type)
+      errors.add(
+        :content_type,
+        "'#{content_type}' not valid with value_type '#{value_type}'"
+      )
+    end
+  end
 
   # Note that we essentially copied this method over from nested question
   def self.update_all_exactly!(content_hashes)
@@ -56,6 +95,27 @@ class CardContent < ActiveRecord::Base
       content.update!(hash)
       update_nested!(child_hashes, content.id, idents)
       content
+    end
+  end
+
+  def to_xml(options = {})
+    attrs = {
+      'content-type' => content_type,
+      'value-type' => value_type,
+      'visible-with-parent-answer' => visible_with_parent_answer
+    }.compact
+    setup_builder(options).tag!('content', attrs) do |xml|
+      safe_dump_text(xml, 'placeholder', placeholder) if placeholder.present?
+      safe_dump_text(xml, 'text', text) if text.present?
+      safe_dump_text(xml, 'label', label) if label.present?
+      if possible_values.present?
+        possible_values.each do |item|
+          xml.tag!('possible-value', label: item['label'], value: item['value'])
+        end
+      end
+      children.each do |child|
+        child.to_xml(builder: xml, skip_instruct: true)
+      end
     end
   end
 end