+++ app/serializers/card_content_serializer.rb
@@ -1,7 +1,23 @@
 class CardContentSerializer < ActiveModel::Serializer
-  attributes :id, :ident, :text, :value_type, :content_type, :order
+  attributes :id,
+             :content_type,
+             :ident,
+             :label,
+             :order,
+             :placeholder,
+             :possible_values,
+             :text,
+             :value_type,
+             # when visible_with_parent_answer is set,
+             # if the parent's answer is equal to this value
+             # then render this content's children
+             :visible_with_parent_answer
 
-  has_many :children, embed: :ids, include: true, root: :card_contents, key: :unsorted_child_ids
+  has_many :children,
+           embed: :ids,
+           include: true,
+           root: :card_contents,
+           key: :unsorted_child_ids
 
   def order
     object.lft