+++ client/app/pods/components/card-content/radio/component.js
@@ -1,0 +1,33 @@
+import Ember from 'ember';
+import { PropTypes } from 'ember-prop-types';
+
+export default Ember.Component.extend({
+  classNames: ['card-content-radio'],
+
+  propTypes: {
+    answer: PropTypes.EmberObject.isRequired,
+    content: PropTypes.EmberObject.isRequired,
+    disabled: PropTypes.bool,
+    owner: PropTypes.EmberObject.isRequired,
+    preview: PropTypes.bool
+  },
+
+  init() {
+    this._super(...arguments);
+
+    Ember.assert(
+      `the content must define an array of possibleValues
+      that contains at least one object with the shape { label, value } `,
+      Ember.isPresent(this.get('content.possibleValues'))
+    );
+  },
+
+  actions: {
+    valueChanged(newVal) {
+      let action = this.get('valueChanged');
+      if (action) {
+        action(newVal);
+      }
+    }
+  }
+});