+++ CHANGELOG.md
@@ -2,29 +2,53 @@
 All notable changes to this project will be documented in this file. Follow
 guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 
-## ## [x.x.x] - {yyyy-mm-dd}
+## ## [1.40.1] - {2017-04-07}
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+Invite Reviewer Card invitation row now shows the correct datetime when in a completed state
+### Security
+
+## ## [1.40.0] - {2017-04-06}
 ### Added
 * Inject the Git Commit ID in <meta> tag for easier debugging
 * Similarity Check Task Card (UI Only)
 ### Changed
 * Edited CSS (color and hover features) for manuscript list on dashboard to be
+* Revise Manuscript card changed to Response to Reviewers, it's attachments are now owned by decisions.
 more user friendly.
+* Updated mathjax cdn to use cloudflare.
+* The confirmation dialogues for removing cards from a paper and a manuscript manager template are
+now clearer about what's going to occur
 ### Deprecated
 ### Removed
 ### Fixed
 ### Security
 
-## ## [1.39.0] - {2017-03-16}
+## ## [1.39.1] - {2017-03-30}
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+* Funders can now be added and updated
+### Security
+
+## ## [1.39.0] - {2017-03-29}
 ### Added
 * The footer discussion cannot be seen by Reviewers and other external users in
   individual card view
 * Billing staff now have more viewing access than just to the Billing card
 * Users can now save lists on Safari for Ad Hoc cards
-* Reviewer Report status available on invitation list 
+* Reviewer Report status available on invitation list
 ### Changed
+* File name for PDF Downloads have been updated to follow the format "[short doi] - [author last name] - [version].pdf"
 ### Deprecated
 ### Removed
 ### Fixed
+* Fixed bug that always returned the latest version of a manuscript regardless of the version that was requested.
 ### Security
 
 ## ## [1.38.0] - {2017-03-09}